Add & update dependencies through `vcpkg` and `find_package`, fix warnings/msvc errors & include issues (#22)
* Remove setting compiler directly in CMakeLists.txt

* Remove setting library paths from environment variables in engine/CMakeLists.txt

* Fix erroneous declaration of `rbe_mem_allocate` - it used different argument type in declaration

* Made library paths uniform

* Add missing include

* Fix paths in Makefile

* Fix paths for editor build
Add separate flags for MSVC compiler

* Add missing glad dependency
Initial vcpkg implementation

* Add member field to `Audio` to avoid compilation error

* Change C std to 11

* Fix narrowing conversion warning

* Update stb_image.h
Fix some warnings

* Change VLA to use malloc

* Use `io.h` insstead of `unistd.h` on msvc

* Fix linking against winsock

* Update C version

* Fix deprecation warnings

* Clean up from merge.

* Fixing missing TODO comment.

* Started working on rough collisions.

* Adding colliders to scene.

* Change `strcpy_s` and such back to standard form
Add msvc deprecation warning disable for `strcpy` and such

* Fix some more `strcpy_s` and VLA

* Fixing sprite shader issues.

* Added iterator for string hash map.

* Updated Makefile.

* CLean.

* Adding device id to input action.

* More device id.

* Starting on gamepad input logic.

* Updating version.

* Cleaning up from previous commits.

* Add msvc warning disable for `fopen`

* Add missing files to cmake

+++ engine/src/core/memory/rbe_mem.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdlib.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -13,7 +15,7 @@ rbe_mem_allocate(Size)
 #define RBE_MEM_FREE(Memory)             \
 rbe_mem_free(Memory)
 
-void* rbe_mem_allocate(unsigned int size);
+void* rbe_mem_allocate(size_t size);
 void rbe_mem_free(void* memory);
 
 #ifdef __cplusplus
